#1538 Fixed Dialog not showing complex widgets. Also add dismissDialog
