+++ modules/ensemble/lib/screen_controller.dart
@@ -242,90 +242,6 @@ class ScreenController {
           executeActionWithScope(context, scopeManager, action.onModalDismiss!);
         });
       }
-    } else if (action is ShowDialogAction) {
-      Widget widget = scopeManager.buildWidgetFromDefinition(action.body);
-
-      // get styles. TODO: make bindable
-      Map<String, dynamic> dialogStyles = {};
-      action.options?.forEach((key, value) {
-        dialogStyles[key] = scopeManager.dataContext.eval(value);
-      });
-
-      bool useDefaultStyle = dialogStyles['style'] != 'none';
-      BuildContext? dialogContext;
-
-      showGeneralDialog(
-          useRootNavigator: false,
-          // use inner-most MaterialApp (our App) as root so theming is ours
-          context: context,
-          barrierDismissible: true,
-          barrierLabel: "Barrier",
-          barrierColor: Colors.black54,
-          // this has some transparency so the bottom shown through
-
-          pageBuilder: (context, animation, secondaryAnimation) {
-            // save a reference to the builder's context so we can close it programmatically
-            dialogContext = context;
-            scopeManager.openedDialogs.add(dialogContext!);
-
-            return Align(
-                alignment: Alignment(
-                    Utils.getDouble(dialogStyles['horizontalOffset'],
-                        min: -1, max: 1, fallback: 0),
-                    Utils.getDouble(dialogStyles['verticalOffset'],
-                        min: -1, max: 1, fallback: 0)),
-                child: Material(
-                    color: Colors.transparent,
-                    child: ConstrainedBox(
-                        constraints: BoxConstraints(
-                            minWidth: Utils.getDouble(dialogStyles['minWidth'],
-                                fallback: 0),
-                            maxWidth: Utils.getDouble(dialogStyles['maxWidth'],
-                                fallback: double.infinity),
-                            minHeight: Utils.getDouble(
-                                dialogStyles['minHeight'],
-                                fallback: 0),
-                            maxHeight: Utils.getDouble(
-                                dialogStyles['maxHeight'],
-                                fallback: double.infinity)),
-                        child: Container(
-                            decoration: useDefaultStyle
-                                ? const BoxDecoration(
-                                    color: Colors.white,
-                                    borderRadius:
-                                        BorderRadius.all(Radius.circular(5)),
-                                    boxShadow: <BoxShadow>[
-                                        BoxShadow(
-                                          color: Colors.white38,
-                                          blurRadius: 5,
-                                          offset: Offset(0, 0),
-                                        )
-                                      ])
-                                : null,
-                            margin: useDefaultStyle
-                                ? const EdgeInsets.all(20)
-                                : null,
-                            padding: useDefaultStyle
-                                ? const EdgeInsets.all(20)
-                                : null,
-                            child: SingleChildScrollView(
-                              child: widget,
-                            )))));
-          }).then((value) {
-        // remove the dialog context since we are closing them
-        scopeManager.openedDialogs.remove(dialogContext);
-
-        // callback when dialog is dismissed
-        if (action.onDialogDismiss != null) {
-          executeActionWithScope(
-              context, scopeManager, action.onDialogDismiss!);
-        }
-      });
-    } else if (action is CloseAllDialogsAction) {
-      for (var dialogContext in scopeManager.openedDialogs) {
-        Navigator.pop(dialogContext);
-      }
-      scopeManager.openedDialogs.clear();
     } else if (action is PlaidLinkAction) {
       final linkToken = action.getLinkToken(scopeManager.dataContext).trim();
       if (linkToken.isNotEmpty) {
